+++ src/test/scala/com/miguno/kafkastorm/integration/IntegrationSuite.scala
@@ -1,9 +1,11 @@
 package com.miguno.kafkastorm.integration
 
+import com.miguno.kafkastorm.spark.KafkaSparkStreamingSpec
 import org.scalatest.Stepwise
 
 class IntegrationSuite extends Stepwise(
   new KafkaSpec,
   new StormSpec,
-  new KafkaStormSpec
+  new KafkaStormSpec,
+  new KafkaSparkStreamingSpec
 )